aggiustato carrello (app dopo ordine si chiude senza dare errori)

+++ justorder_framework/src/main/java/it/polito/justorder_framework/common_activities/OrderDetails.java
@@ -29,7 +29,7 @@ public class OrderDetails extends AbstractViewerWithImagePickerActivityAndToolba
     private User user;
     private TextView orderIdTextField, userTextField, addressTextField, priceTextField, timestampTextField, riderTextField, productsTextField;
     private String productString;
-    private Map<Product, Integer> products = new HashMap<>();
+    private Map<Product,Double> products = new HashMap<>();
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -81,7 +81,7 @@ public class OrderDetails extends AbstractViewerWithImagePickerActivityAndToolba
                 }));
             }
 
-            for(Map.Entry<String, Integer> entry : this.order.getProducts().entrySet()){
+            for(Map.Entry<String,Double> entry : this.order.getProducts().entrySet()){
                 Database.INSTANCE.getProducts().get(entry.getKey(), (product -> {
                     this.products.put(product, entry.getValue());
                     this.reloadViews();
@@ -121,7 +121,7 @@ public class OrderDetails extends AbstractViewerWithImagePickerActivityAndToolba
     private String createProductString() {
 
         String productString = "";
-        for(Map.Entry<Product, Integer> entry : products.entrySet()){
+        for(Map.Entry<Product, Double> entry : products.entrySet()){
             productString += entry.getKey().getName() + ";\n";
         }
